refactor: moved tests in test folder
